step1 완성
